app/requirements.txt
fastapi[standard]
passlib[bcrypt]
sqlmodel
PyJWT
python-dotenv